requirements.txt
Flask==3.0.0
flask-cors==6.0.1
python-dotenv==1.0.1
google-genai==1.70.0
gunicorn==21.2.0